feat: nextcloud set samba admin as nextcloud admin

+++ docker_compose/nextcloud/nextcloud/root/etc/cont-init.d/03-config_extra.sh
@@ -4,4 +4,6 @@ if [ "$SIDECAR_CRON" = "1" ] || [ "$SIDECAR_PREVIEWGEN" = "1" ] || [ "$SIDECAR_N
   exit 0
 fi
 
-rm /var/www/config/autoconfig.php
+if [ -f /var/www/config/autoconfig.php ] ; then
+  rm /var/www/config/autoconfig.php
+fi
